Support for host file bindings

+++ src/syscalls/syscalls.h
@@ -56,7 +56,7 @@ procpath(struct TuxProc* p, lfiptr_t pathp)
 
 ssize_t sys_write(struct TuxProc* p, int fd, lfiptr_t bufp, size_t size);
 
-uintptr_t sys_exit(struct TuxThread* p, int val);
+uintptr_t sys_exit(struct TuxThread* p, uint64_t val);
 
 uintptr_t sys_brk(struct TuxProc* p, lfiptr_t addr);
 
@@ -74,7 +74,7 @@ typedef int tux_clockid_t;
 
 int sys_clock_gettime(struct TuxProc* p, tux_clockid_t clockid, lfiptr_t tp);
 
-uintptr_t sys_exit_group(struct TuxThread* p, int code);
+uintptr_t sys_exit_group(struct TuxThread* p, uint64_t code);
 
 int sys_ioctl(struct TuxProc* p, int fd, unsigned long request, uintptr_t va0,
         uintptr_t va1, uintptr_t va2, uintptr_t va3);
